Refactor database configuration and dependencies for SQLite
- Changed database dialect from PostgreSQL to SQLite in drizzle.config.ts.
- Updated database URL to point to a local SQLite file.
- Removed PostgreSQL dependencies from package.json and bun.lock.
- Added better-sqlite3 as a trusted dependency.
- Deleted obsolete migration files related to PostgreSQL schema.

+++ src/db/repositories/feature-message.repository.ts
@@ -54,7 +54,7 @@ export class FeatureMessageRepository {
     const counts = await db
       .select({
         featureId: schema.featureMessages.featureId,
-        count: sql<number>`count(*)::int`,
+        count: sql<number>`count(*)`,
       })
       .from(schema.featureMessages)
       .where(inArray(schema.featureMessages.featureId, featureIds))
